revisions to reflect new understanding of ns, require, defrecord
The rule is that you need a require *in the ns* decl and then you
can use record constructors in the form "->BlahBlah".
import is only need for "BlahBlah." constructors, and I couldn't
get it to work with the namespace switching.

Also did some renaming, and distinguished between the config data
class and the record constructor, which previously shared names,
incorrectly.