inc'd initial-energy to 10, birth-threshold to 20
I originally had initial-energy at 10, but reduced it to 5
partly because 10 was very close to the birth-threshold of 15,
and the cost of birth was 5.

However, initial-energy 5 is hard for k-snipes.  It makes it
difficult to survive long enough while learning.  It might work
if I adjusted the learning speed.  Instead I'm just increasing
initial energy, but also increasing birth-threshold so
it's not so close to initial energy.