moved reproduction before eating in next-popenv
this changes the dynamics slightly, but seems ok.  it still
puts both eating and reproduction before death, and allows
eating to take place after movement later on so that we can
see snipes sitting on their future food.

I made the change because in the multifield branch, this
will be more convenient and I wanted to test it in an known
good setup and also maintain as much parallelism between
master (might be renamed soon) and multifield.