major bugfix in next-popenv: fixed problem with mushs not updating
Problem was that there were some RHS's in the big let that used the
original passed-in mush fields and not ones updated in previous lines.
So the problem wasn't in the GUI, as I thought.  It was in the
underlying simulations.  Snipes were eating the same mushrooms
in the same places over and over again.