Revert "broken with-open csv writing" (this will never work)
This reverts commit 4a573b29a6f41a4f6492d015335c339f1691dcd8.

The commit tried to use with-open for the csv writing.  I believe it's
not going to work because Mason will schedule my reify's and then run
them outside of the with-open, which will have closed the file.  I need
to explicitly open it by hand, and close it from inside the schedule
reify's.