removed max-ticks, renamed -f to -F
The functionality of max-ticks can be done using MASON's -for (or -until),
and now I have a custom finish() function that will automatically get
called by SimState's finish() whenever MASON ends a simlation.

MASON's -for was being interepreted as our -f plus the string "or".
I don't know why there's a clash here but not with other options (afaics).