replaced old snipe next-id with mush's gensym-based version
The point of the special snipe version was to allow id numbers
to also make it apparent how many snipes had lived.  This is
rarely useful.  It only works on the first run, and the only
time I used it was in experiments in the repl, I think.
It serves no purpose in the GUI, and at present I don't record
ids in output date files.  I just don't see the point.  I should
use whatever system is most efficient.  Not sure if the gensym
version is best, but it's simplest to use the same system for
both mushrooms and snipes.